align indexed TLSU (MGATHER/MSCATTER) with v0.58.6 element semantics (#130)
* TileOP: align indexed TLSU with v0.58.6 element semantics

* docs: fix MGATHER_CAS example carrier dtype (U32 per CAS contract)

The doc-check compiles every example; the MGATHER_CAS page still showed
a float transfer carrier after 8e4d66e enforced the CAS contract
(U16/U32/U64 only). Align the example with the contract.

---------

+++ test/tileop_api/src/MGatherCas.cpp
@@ -1,10 +1,10 @@
-// MGATHER_CAS: atomic compare-and-swap at byte displacements
-// (PTO ISA 0.58.3 TLSU function 8; canonical BSTART.TLSU MGATHER.CAS).
+// MGATHER_CAS: atomic compare-and-swap at logical element indices
+// (PTO ISA 0.58.6 TLSU function 8; canonical BSTART.TLSU MGATHER.CAS).
 #include <common/pto_tileop.hpp>
 
 using namespace pto;
 
-using D = Tile<Location::Vec, float, 8, 256, BLayout::RowMajor>;
+using D = Tile<Location::Vec, uint32_t, 8, 256, BLayout::RowMajor>;
 using Idx16 = Tile<Location::Vec, int16_t, 8, 256, BLayout::RowMajor>;
 using Idx4 = Tile<Location::Vec, __int4x2, 8, 256, BLayout::RowMajor>;
